Split hooks.ts.
